Float/Double issue still

+++ include/messages/kinect_messages.h
@@ -287,7 +287,7 @@ public:
 class KinectBodyLeanMessage
 {
 public:
-	double x, y;
+	float x, y;
 
 	// ====================================================================================================
 	KinectBodyLeanMessage()
